feat(cli): show descriptor discrepancies in non-interactive mode
