+++ packages/cli/src/main.ts
@@ -24,10 +24,9 @@ import {
 } from "./io"
 import {
   checkDescriptorsForDiscrepancies,
+  showDiscrepancies,
   synchronizeDescriptors,
 } from "./sync"
-import chalk from "chalk"
-import asTable from "as-table"
 import { blowupMetadata } from "./testing"
 
 const ProgramArgs = z.object({
@@ -69,12 +68,16 @@ const descriptorMetadata = await readDescriptors({
 if (descriptorMetadata && !options.interactive) {
   for (const key of Object.keys(descriptorMetadata)) {
     const data = await Data.fromSavedDescriptors(descriptorMetadata[key])
-    if (data.isInitialized && data.outputFolder) {
-      if (options.sync) {
-        const discrepancies = checkDescriptorsForDiscrepancies(data)
-        synchronizeDescriptors(data, discrepancies)
-      }
 
+    if (data.isInitialized && data.outputFolder) {
+      const discrepancies = checkDescriptorsForDiscrepancies(data)
+
+      if (options.sync) {
+        synchronizeDescriptors(data, discrepancies)
+      } else if (discrepancies.length > 0) {
+        showDiscrepancies(discrepancies)
+        process.exit(1)
+      }
       if (!options.interactive) {
         await outputCodegen(data, data.outputFolder, key)
       }
@@ -329,65 +332,7 @@ while (!exit) {
     case SYNC: {
       const discrepancies = checkDescriptorsForDiscrepancies(data)
 
-      const mapDiscrepancy = ({
-        pallet,
-        name,
-        oldChecksum,
-        newChecksum,
-      }: (typeof discrepancies)[number]) => ({
-        Pallet: pallet,
-        Name: name,
-        "Old Checksum":
-          oldChecksum === null ? chalk.red(oldChecksum) : oldChecksum,
-        "New Checksum":
-          newChecksum === null ? chalk.red(newChecksum) : newChecksum,
-      })
-
-      console.log("-------- Constant Discrepancies --------")
-      console.log(
-        asTable(
-          discrepancies
-            .filter(({ type }) => type === "constant")
-            .map(mapDiscrepancy),
-        ),
-      )
-      console.log("")
-      console.log("-------- Storage Discrepancies --------")
-      console.log(
-        asTable(
-          discrepancies
-            .filter(({ type }) => type === "storage")
-            .map(mapDiscrepancy),
-        ),
-      )
-      console.log("")
-      console.log("-------- Event Discrepancies --------")
-      console.log(
-        asTable(
-          discrepancies
-            .filter(({ type }) => type === "event")
-            .map(mapDiscrepancy),
-        ),
-      )
-      console.log("")
-      console.log("-------- Error Discrepancies --------")
-      console.log(
-        asTable(
-          discrepancies
-            .filter(({ type }) => type === "error")
-            .map(mapDiscrepancy),
-        ),
-      )
-      console.log("")
-      console.log("-------- Extrinsic Discrepancies --------")
-      console.log(
-        asTable(
-          discrepancies
-            .filter(({ type }) => type === "extrinsic")
-            .map(mapDiscrepancy),
-        ),
-      )
-      console.log("")
+      showDiscrepancies(discrepancies)
 
       if (discrepancies.length > 0) {
         const accept = await confirm({ message: "Accept changes" })
